requirements.txt
ultralytics
torch
torchvision
numpy
opencv-python
Pillow
fastapi
uvicorn
python-multipart